Major refactoring
I rewrote most of the content of 'deployVmsByList.ps1' into 'utils.ps1',
but in a much more usable manner. The other two powershell scripts simply
consume utils.
Both deploy scripts actually work- though the target datastore is
hardcoded, so the VM being cloned must already be on a host that can reach
that datastore. I will fix this in the future.